Revert "feat(diff): add side-by-side firmware hex diff (#148)"
This reverts commit 523f44ee19ccb29cfa99a748e11b76222eff5da7.
