--- .trellis/tasks/archive/2026-08/08-01-hex-diff-view/implement.jsonl
@@ -1,1 +1,0 @@
-{"_example": "Fill with {\"file\": \"<path>\", \"reason\": \"<why>\"}. Put spec/research files only — no code paths. Run `python .trellis/scripts/get_context.py --mode packages` to list available specs. Delete this line once real entries are added."}
