--- .trellis/tasks/archive/2026-08/08-01-hex-diff-view/task.json
@@ -1,26 +1,0 @@
-{
-  "id": "hex-diff-view",
-  "name": "hex-diff-view",
-  "title": "Hex diff view: compare two firmware files",
-  "description": "",
-  "status": "completed",
-  "dev_type": null,
-  "scope": null,
-  "package": null,
-  "priority": "P2",
-  "creator": "deviousprophet",
-  "assignee": "deviousprophet",
-  "createdAt": "2026-08-01",
-  "completedAt": "2026-08-01",
-  "branch": "feat/hex-diff-view",
-  "base_branch": "main",
-  "worktree_path": null,
-  "commit": null,
-  "pr_url": null,
-  "subtasks": [],
-  "children": [],
-  "parent": null,
-  "relatedFiles": [],
-  "notes": "",
-  "meta": {}
-}
