+++ .trellis/workspace/deviousprophet/index.md
@@ -8,7 +8,7 @@
 
 <!-- @@@auto:current-status -->
 - **Active File**: `journal-1.md`
-- **Total Sessions**: 3
+- **Total Sessions**: 2
 - **Last Active**: 2026-08-01
 <!-- @@@/auto:current-status -->
 
@@ -19,7 +19,7 @@
 <!-- @@@auto:active-documents -->
 | File | Lines | Status |
 |------|-------|--------|
-| `journal-1.md` | ~106 | Active |
+| `journal-1.md` | ~50 | Active |
 <!-- @@@/auto:active-documents -->
 
 ---
@@ -29,7 +29,6 @@
 <!-- @@@auto:session-history -->
 | # | Date | Title | Commits | Branch |
 |---|------|-------|---------|--------|
-| 3 | 2026-08-01 | Hex diff view complete: lazy-window model, reusable components, review cleanup | `768e285`, `9e8ce5d`, `b51bd29`, `ea1c94c`, `026c090`, `a1ff249`, `7961483`, `b3168f0`, `a692e48` | `feat/hex-diff-view` |
 | 2 | 2026-08-01 | Fix gap copy zero + webview test suite nesting repair | `dc02597` | `fix/gap-copy-zero` |
 | 1 | 2026-07-30 | Bootstrap guidelines + finish-work | `b9ced7d`, `b0a8e58` | `main` |
 <!-- @@@/auto:session-history -->
