+++ src/webview/styles/struct.css
@@ -126,19 +126,12 @@
     font-size: 11px; cursor: pointer; padding: 0; line-height: 1.4;
     transition: border-color .1s, color .1s, background .1s;
 }
-.sfe-ptr-btn:hover:not(:disabled),
-.sfe-arr-toggle:hover,
-.sfe-bit-btn:hover:not(:disabled) { border-color: var(--high-color); color: var(--high-color); }
-.sfe-ptr-btn.active,
-.sfe-arr-toggle.active,
-.sfe-bit-btn.sfe-bit-btn-on { border-style: solid; border-color: var(--high-color);
-                              color: var(--high-color); background: rgba(79,195,247,.1); }
-.sfe-ptr-btn.active:hover,
-.sfe-arr-toggle.active:hover,
-.sfe-bit-btn.sfe-bit-btn-on:hover { border-color: var(--err); color: var(--err);
-                                    background: rgba(255,80,80,.08); }
-.sfe-ptr-btn:disabled,
-.sfe-bit-btn:disabled { opacity: .35; cursor: not-allowed; }
+.sfe-ptr-btn:hover:not(:disabled) { border-color: var(--high-color); color: var(--high-color); }
+.sfe-ptr-btn.active { border-style: solid; border-color: var(--high-color);
+                      color: var(--high-color); background: rgba(79,195,247,.1); }
+.sfe-ptr-btn.active:hover { border-color: var(--err); color: var(--err);
+                            background: rgba(255,80,80,.08); }
+.sfe-ptr-btn:disabled { opacity: .35; cursor: not-allowed; }
 
 /* Array cell */
 .sfe-arr-cell {
@@ -151,6 +144,11 @@
     white-space: nowrap; flex-shrink: 0; text-align: center;
     transition: border-color .1s, color .1s, background .1s;
 }
+.sfe-arr-toggle:hover        { border-color: var(--high-color); color: var(--high-color); }
+.sfe-arr-toggle.active       { border-style: solid; border-color: var(--high-color);
+                                color: var(--high-color); background: rgba(79,195,247,.1); }
+.sfe-arr-toggle.active:hover { border-color: var(--err); color: var(--err);
+                                background: rgba(255,80,80,.08); }
 .sfe-count-inp,
 .sfe-bf-child-width {
     width: 32px; text-align: center;
@@ -165,13 +163,11 @@
 .sfe-count-inp:focus { border-color: var(--focus-bdr); }
 .sfe-arr-cell.is-array .sfe-count-inp { display: block; }
 
-.sfe-del-btn,
-.sfe-bf-del-child {
+.sfe-del-btn {
     background: transparent; border: none; color: var(--addr-fg);
     font-size: 10px; cursor: pointer; padding: 0; line-height: 1; text-align: center;
 }
-.sfe-del-btn:hover,
-.sfe-bf-del-child:hover { color: var(--err); }
+.sfe-del-btn:hover { color: var(--err); }
 .sfe-del-placeholder { display: block; width: 18px; }
 
 /* -- Bit-field toggle button (:N) ------------------------------- */
@@ -182,6 +178,12 @@
     font-size: 9px; cursor: pointer; padding: 1px 1px; line-height: 1.4;
     white-space: nowrap; transition: border-color .1s, color .1s, background .1s;
 }
+.sfe-bit-btn:hover:not(:disabled)        { border-color: var(--high-color); color: var(--high-color); }
+.sfe-bit-btn.sfe-bit-btn-on              { border-style: solid; border-color: var(--high-color);
+                                            color: var(--high-color); background: rgba(79,195,247,.1); }
+.sfe-bit-btn.sfe-bit-btn-on:hover        { border-color: var(--err); color: var(--err);
+                                            background: rgba(255,80,80,.08); }
+.sfe-bit-btn:disabled                    { opacity: .35; cursor: not-allowed; }
 
 .si-f-ptr-note {
     color: var(--addr-fg);
@@ -228,6 +230,11 @@
 .sfe-bf-child-move {
     display: flex; align-items: center; gap: 1px;
 }
+.sfe-bf-del-child {
+    background: transparent; border: none; color: var(--addr-fg);
+    font-size: 10px; cursor: pointer; padding: 0; line-height: 1; text-align: center;
+}
+.sfe-bf-del-child:hover { color: var(--err); }
 .sfe-bf-add-child {
     background: transparent; border: 1px dashed var(--border);
     border-radius: 2px; color: var(--addr-fg); font-size: 9px;
@@ -418,8 +425,7 @@
     line-height: 1; align-self: flex-start; margin-top: 2px;
 }
 
-.si-card-info,
-.integrity-card-info {
+.si-card-info {
     flex: 1; min-width: 0;
     display: flex; flex-direction: column; gap: 2px;
 }
@@ -587,7 +593,7 @@
     border-bottom: 1px dotted rgba(255,255,255,.2);
     margin-bottom: 2px; align-self: flex-end;
 }
-.si-f-pri,
+.si-f-pri { font-size: 10px; color: var(--high-color); font-family: var(--font-editor); white-space: nowrap; flex-shrink: 0; }
 .si-f-val { font-size: 10px; color: var(--high-color); font-family: var(--font-editor); white-space: nowrap; flex-shrink: 0; }
 .si-f-val[data-val-type="hex"]   { color: var(--val-hex-color); }
 .si-f-val[data-val-type="dec"]   { color: var(--val-dec-color); }
