--- src/webview/ui-components/search-bar/searchBarComponent.css
@@ -1,21 +1,0 @@
-/* Reusable search bar (mode select, endian toggle, input, nav buttons, count).
-   Self-contained: the host only needs base.css (shared search-box/nav/chrome
-   rules live there) + this file. */
-
-/* Endian toggle: segmented pill (Auto / LE / BE) like the single view */
-.search-endian-toggle {
-    flex-shrink: 0;
-    display: inline-flex; align-items: center; gap: 1px;
-    background: rgba(0,0,0,.22); border-radius: 4px; padding: 2px;
-}
-.search-endian-toggle button {
-    min-width: 42px; height: 22px; font-size: 10px;
-    border: none; border-radius: 3px;
-    background: transparent; color: var(--fg);
-    font-family: var(--font-ui); cursor: pointer;
-}
-.search-endian-toggle button:hover:not(.active) { background: rgba(255,255,255,.07); }
-.search-endian-toggle button.active { background: var(--btn-bg); color: var(--btn-fg); }
-
-/* button sizes mirror the single view exactly */
-.search-btn        { width: 24px; height: 24px; padding: 0; font-size: 12px; line-height: 1; }
