+++ src/webview/webviewMessageModel.ts
@@ -73,11 +73,6 @@ const MODEL_APPLIERS: ModelAppliers = {
     scriptResult: applyPassiveMessage,
     scriptOutput: applyPassiveMessage,
     activateScriptsTab: applyPassiveMessage,
-    diffInit: applyPassiveMessage,
-    diffUpdate: applyPassiveMessage,
-    diffProgress: applyPassiveMessage,
-    diffSwap: applyPassiveMessage,
-    diffSearch: applyPassiveMessage,
 };
 
 function applyPassiveMessage(): WebviewModelUpdate { return { invalidations: {} }; }
