+++ src/webviewProtocol.ts
@@ -1,8 +1,6 @@
 import type { CopyCommand } from './core/byte-tools/copyCommand';
-import type { DiffMeta } from './core/diff';
 import type { HexScopeFormat } from './core/document';
 import type { IntegrityCheckSet, IntegrityProfile } from './core/integrity';
-import type { SearchEndianness, SearchMode } from './core/types';
 import type { SegmentLabel, SerializedRecord, StructDef, StructPin, WireParseResult } from './core/types';
 
 export const RECORD_PAGE_SIZE = 512;
@@ -43,12 +41,7 @@ export type ProviderToWebviewMessage =
     | { type: 'scriptInfo'; trusted: boolean; scripts: Array<{ name: string; filePath: string; capabilities: string[] }> }
     | { type: 'scriptResult'; scriptPath: string; result: { results: Array<{ label: string; value: string }>; log: string[] } | null; error: string; errorType?: 'compile' | 'runtime' | 'timeout' | 'cancel'; pendingWriteCount: number }
     | { type: 'scriptOutput'; scriptPath: string; text: string }
-    | { type: 'activateScriptsTab' }
-    | { type: 'diffInit'; generation: number; a: WireParseResult; b: WireParseResult; meta: DiffMeta; aLabel: string; bLabel: string; aFormat: HexScopeFormat; bFormat: HexScopeFormat; aError: string | null; bError: string | null }
-    | { type: 'diffUpdate'; generation: number; a: WireParseResult; b: WireParseResult; meta: DiffMeta; aError: string | null; bError: string | null }
-    | { type: 'diffProgress'; generation: number; stage: 'read' | 'parse' | 'build' | 'transfer'; completed: number; total: number }
-    | { type: 'diffSwap'; generation: number; swapped: boolean }
-    | { type: 'diffSearch'; generation: number; query: string; matches: number[]; done: boolean };
+    | { type: 'activateScriptsTab' };
 
 export type WebviewToProviderMessage =
     | { type: 'ready' }
@@ -72,10 +65,7 @@ export type WebviewToProviderMessage =
     | { type: 'viewInNormalEditor' }
     | { type: 'requestScriptList' }
     | { type: 'runScript'; scriptPath: string; generation: number; selectionRange?: { start: number; end: number } }
-    | { type: 'cancelScript'; scriptPath: string }
-    | { type: 'diffReady' }
-    | { type: 'diffSwapRequest' }
-    | { type: 'diffSearchRequest'; generation: number; query: string; mode: SearchMode; endianness: SearchEndianness };
+    | { type: 'cancelScript'; scriptPath: string };
 
 export function messageType(message: unknown): string | undefined {
     return typeof (message as { type?: unknown })?.type === 'string'
